Remove Schm_hzdadw_L and associated routine hzd_momentum.F90 Introduce Schm_POSO(default=2) to control precicion order for 3D spacial operators Introduce delQ.F90 in matvec and bac Fix findtopo

+++ src/gemdyn/src/base/nest_bcs.F90
@@ -22,7 +22,7 @@
       implicit none
 
       integer, intent(in) :: Minx, Maxx, Miny, Maxy, Nk
-      real, dimension(Minx:Maxx,Miny:Maxy,Nk), intent(out) :: &
+      real(kind=REAL64), dimension(Minx:Maxx,Miny:Maxy,Nk), intent(out) :: &
                                                F_rhsu, F_rhsv
       real(kind=REAL64), intent(IN) :: F_dt_8
       real(kind=REAL64) :: invT_8
